+++ PhoneBook.Cactus/Program.cs
@@ -1,4 +1,4 @@
-﻿using PhoneBook.Cactus;
+﻿using PhoneBook.Cactus.Service;
 using Spectre.Console;
 enum MenuOptions
 {
